+++ platformio_core_defs.ini
@@ -38,7 +38,7 @@ build_flags               = -D NDEBUG
                             -DVTABLES_IN_FLASH
                             -DPUYA_SUPPORT=1
                             -fno-strict-aliasing
-                            -I$PROJECT_DIR/include
+                            -I$PROJECT_DIR/src/include
                             -include "ESPEasy_config.h"
 
 lib_ignore                = ESP32_ping
@@ -60,7 +60,7 @@ build_flags               = -DNDEBUG
                             -DCORE_POST_2_5_0
                             -fno-strict-aliasing
                             -DLIBRARIES_NO_LOG=1
-                            -I$PROJECT_DIR/include
+                            -I$PROJECT_DIR/src/include
                             -include "ESPEasy_config.h"
 lib_ignore                = ${esp82xx_defaults.lib_ignore}
 
@@ -235,7 +235,7 @@ platform_packages           = framework-arduinoespressif32 @ https://github.com/
 build_flags                 = -DESP32_STAGE
                               -DMUSTFIX_CLIENT_TIMEOUT_IN_SECONDS
                               -DLIBRARIES_NO_LOG=1
-                              -I$PROJECT_DIR/include
+                              -I$PROJECT_DIR/src/include
                               -include "sdkconfig.h"
                               -include "ESPEasy_config.h"
 
@@ -245,6 +245,6 @@ platform_packages           =
 build_flags                 = -DESP32_STAGE
                               -DMUSTFIX_CLIENT_TIMEOUT_IN_SECONDS
                               -DLIBRARIES_NO_LOG=1
-                              -I$PROJECT_DIR/include
+                              -I$PROJECT_DIR/src/include
                               -include "sdkconfig.h"
                               -include "ESPEasy_config.h"
