+++ src/ESPEasy_common.h
@@ -16,7 +16,7 @@
 
 #include <Arduino.h>
 // User configuration
-#include "../include/ESPEasy_config.h"
+#include "include/ESPEasy_config.h"
 #include "./src/CustomBuild/ESPEasyDefaults.h"
